refactor(tui): extract model-domain helpers to model-choice
