+++ apps/kimi-code/src/tui/components/dialogs/model-selector.ts
@@ -8,42 +8,20 @@ import {
 } from '@earendil-works/pi-tui';
 import chalk from 'chalk';
 
-import { DEFAULT_OAUTH_PROVIDER_NAME, PRODUCT_NAME } from '#/constant/app';
 import type { ColorPalette } from '#/tui/theme/colors';
 import { SearchableList } from '#/tui/utils/searchable-list';
 
-import type { ChoiceOption } from './choice-picker';
-
-type ThinkingAvailability = 'toggle' | 'always-on' | 'unsupported';
-
-interface ModelChoice {
-  readonly alias: string;
-  readonly model: ModelAlias;
-  readonly label: string;
-}
+import {
+  createModelChoices,
+  effectiveThinking,
+  renderThinkingControl,
+  thinkingAvailability,
+  type ModelChoice,
+} from './model-choice';
 
 export interface ModelSelection {
   readonly alias: string;
   readonly thinking: boolean;
-}
-
-export function modelDisplayName(alias: string, model: ModelAlias | undefined): string {
-  return model?.displayName ?? model?.model ?? alias;
-}
-
-export function providerDisplayName(provider: string): string {
-  if (provider === DEFAULT_OAUTH_PROVIDER_NAME) return PRODUCT_NAME;
-  if (provider.startsWith('managed:')) return provider.slice('managed:'.length);
-  return provider;
-}
-
-export function createModelChoiceOptions(
-  models: Record<string, ModelAlias>,
-): readonly ChoiceOption[] {
-  return Object.entries(models).map(([alias, cfg]) => ({
-    value: alias,
-    label: `${modelDisplayName(alias, cfg)} (${providerDisplayName(cfg.provider)})`,
-  }));
 }
 
 export interface ModelSelectorOptions {
@@ -58,28 +36,6 @@ export interface ModelSelectorOptions {
   readonly pageSize?: number;
   readonly onSelect: (selection: ModelSelection) => void;
   readonly onCancel: () => void;
-}
-
-function createModelChoices(models: Record<string, ModelAlias>): readonly ModelChoice[] {
-  return Object.entries(models).map(([alias, cfg]) => ({
-    alias,
-    model: cfg,
-    label: `${modelDisplayName(alias, cfg)} (${providerDisplayName(cfg.provider)})`,
-  }));
-}
-
-function thinkingAvailability(model: ModelAlias): ThinkingAvailability {
-  const caps = model.capabilities ?? [];
-  if (caps.includes('always_thinking')) return 'always-on';
-  if (caps.includes('thinking')) return 'toggle';
-  return 'unsupported';
-}
-
-function effectiveThinking(model: ModelAlias, thinkingDraft: boolean): boolean {
-  const availability = thinkingAvailability(model);
-  if (availability === 'always-on') return true;
-  if (availability === 'unsupported') return false;
-  return thinkingDraft;
 }
 
 export class ModelSelectorComponent extends Container implements Focusable {
@@ -177,7 +133,7 @@ export class ModelSelectorComponent extends Container implements Focusable {
     lines.push(chalk.hex(colors.textMuted)(' Thinking'));
     const selected = choices[view.selectedIndex];
     if (selected !== undefined) {
-      lines.push(this.renderThinkingControl(selected.model));
+      lines.push(renderThinkingControl(selected.model, this.thinkingDraft, colors));
     }
     lines.push('');
     if (view.page.pageCount > 1) {
@@ -189,22 +145,5 @@ export class ModelSelectorComponent extends Container implements Focusable {
     }
     lines.push(chalk.hex(colors.primary)('─'.repeat(width)));
     return lines.map((line) => truncateToWidth(line, width));
-  }
-
-  private renderThinkingControl(model: ModelAlias): string {
-    const { colors } = this.opts;
-    const segment = (label: string, active: boolean): string =>
-      active
-        ? chalk.hex(colors.primary).bold(`[ ${label} ]`)
-        : chalk.hex(colors.text)(`  ${label}  `);
-
-    const availability = thinkingAvailability(model);
-    if (availability === 'always-on') {
-      return `  ${segment('Always on', true)}`;
-    }
-    if (availability === 'unsupported') {
-      return `  ${segment('Off', true)} ${chalk.hex(colors.textMuted)('unsupported')}`;
-    }
-    return `  ${segment('On', this.thinkingDraft)}  ${segment('Off', !this.thinkingDraft)}`;
   }
 }
